refactor(episode-flow): extract in-episode food diary into hooks and step components (mobile + web)
